renaming files fastsk
