+++ fastsk/src/fastsk.cpp
