+++ fastsk/src/fastsk.hpp
