+++ fastsk/src/kernel.cpp
@@ -1,5 +1,5 @@
 #include "kernel.hpp"
-#include "gakco_core.hpp"
+#include "fastsk.hpp"
 #include "shared.h"
 #include <vector>
 #include <array>
